refactor(wiki/routes): align route filenames with /-/review URL
Previously the files were named `$wikiSlug.dash-review.tsx` but used
explicit `createFileRoute("...$wikiSlug/-/review")` overrides to map to
`/-/review`. TanStack Router's plugin auto-syncs the createFileRoute
path string to match the filename on every dev/build cycle, silently
undoing the override and routing through `/dash-review` instead.

Fix the root cause by using the natural file→URL convention:
- Rename routes + tests to `$wikiSlug.-.review.tsx` (and proposalId variant)
  so the dot-separated segment `-` directly maps to `/-/review`
- Update test imports from `../$wikiSlug.dash-review` to `.-.review`
- Vite config: ignore `__tests__/` and `*.test.*` so the router plugin
  doesn't scan test files and report "does not contain any route piece"

Net effect: file name = URL = `/-/review`. The Round 3 fix that prevented
collision with wiki pages literally named `review/` is preserved without
any string override fighting the plugin. Tests pass (519/519, all 36
suites), build emits 21 `/-/review` references in routeTree.gen.ts.
